requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aioslsk==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==4.0.3
async_upnp_client==0.38.3
attrs==23.2.0
beautifulsoup4==4.13.3
bitarray==3.0.0
bitstring==4.3.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
codebase-digest==0.1.43
colorama==0.4.6
construct==2.10.70
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.15
Django==5.2
djangorestframework==3.16.0
docutils==0.21.2
ecdsa==0.19.0
esp-coredump==1.13.1
esp-idf-monitor==1.6.0
esp_idf_panic_decoder==1.3.0
esptool==4.8.1
Flask==3.1.0
frozenlist==1.4.1
geographiclib==2.0
gerbolyze==3.1.9
gerbonara==1.4.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
intelhex==2.3.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
libusb==1.0.27.post4
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.0.5
mutagen==1.47.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pkg_about==1.2.11
priority==2.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyelftools==0.32
pyftdi==0.56.0
pygdbmi==0.11.0.0
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.2
pyperclip==1.9.0
pyserial==3.5
python-didl-lite==1.4.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
Quart==0.20.0
reedsolo==1.7.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resvg-wasi==0.31.0
rsa==4.9
Rtree==1.3.0
setuptools==77.0.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
svg-flatten-wasi==3.1.9
text-unidecode==1.3
tiktoken==0.8.0
types-aiofiles==23.2.0.20240403
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
voluptuous==0.14.2
wasmtime==29.0.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.9.4
zipp==3.21.0
